+++ src/lib.rs
@@ -25,11 +25,12 @@ pub mod help_content;
 pub mod help_states;
 pub mod help_verbs;
 pub mod io;
-pub mod key_str;
+pub mod keys;
 pub mod mad_skin;
 pub mod matched_string;
 pub mod permissions;
 pub mod screens;
+pub mod selection_type;
 pub mod shell_install;
 pub mod skin;
 pub mod skin_conf;
@@ -45,3 +46,4 @@ pub mod errors;
 pub mod fuzzy_patterns;
 pub mod patterns;
 pub mod regex_patterns;
+