+++ src/selection_type.rs
@@ -1,0 +1,7 @@
+
+#[derive(Debug, PartialEq, Eq, Clone, Copy)]
+pub enum SelectionType {
+    File,
+    Directory,
+    Any,
+}